power up  class added! now players have a chance to slow down the paddles if they made them go too fast!
the power up is on a timer and should come in every 30seconds! it also moves randomly to make it even more...CHALLENGING !

+++ exercises/Final/Paddles.pde
@@ -2,7 +2,7 @@
 class Paddles {
 
   int x;
-  int y;
+  float y;
   float vy = 1;
   float vx;
 
@@ -17,7 +17,7 @@ class Paddles {
   color paddlesColor = color(255);
 
 
-  Paddles( int X, int Y) {
+  Paddles( int X, float Y) {
 
     x = X;
     y = Y;
